move Path into Idea
