+++ backoffice/src/resources/Idea/Path/PathFilterSidebar.js
