+++ backoffice/src/resources/Idea/Path/PathList.js
