+++ backoffice/src/resources/Idea/Path/PathShow.js
