+++ backoffice/src/resources/Idea/Path/PathTitle.js
