+++ frontoffice/src/resources/Idea/Path/PathCard.js
