+++ frontoffice/src/resources/Idea/Path/index.js
