+++ frontoffice/src/resources/index.js
@@ -1,7 +1,9 @@
+// Ideas
+export { default as Path } from './Idea/Path';
+
 // Activities
 export { default as Course } from './Activity/Course';
 export { default as Event } from './Activity/Event';
-export { default as Path } from './Activity/Path';
 
 // Actors
 export { default as Organization } from './Actor/Organization';
